Split darkfield model components
